requirements.txt
uvicorn
fastapi
Pillow
python-multipart